Minor TA adjustments
