+++ modules/Manager.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 from typing import Callable, Iterable
 
 from tqdm import tqdm
@@ -489,7 +490,7 @@ class Manager:
         return None
 
 
-    def report_missing(self, file: 'Path') -> None:
+    def report_missing(self, file: Path) -> None:
         """Report all missing assets for all shows."""
 
         # Serial mode won't have an accurate show list
