+++ modules/cards/TintedGlassTitleCard.py
@@ -1,14 +1,18 @@
 from collections import namedtuple
 from pathlib import Path
-from typing import Literal, Optional
+from typing import TYPE_CHECKING, Literal, Optional
 
 from modules.BaseCardType import BaseCardType, ImageMagickCommands
 from modules.Debug import log
+
+if TYPE_CHECKING:
+    from modules.Font import Font
 
 
 SeriesExtra = Optional
 BoxCoordinates = namedtuple('BoxCoordinates', ('x0', 'y0', 'x1', 'y1'))
 Position = Literal['left', 'center', 'right']
+
 
 class TintedGlassTitleCard(BaseCardType):
     """
@@ -333,7 +337,7 @@ class TintedGlassTitleCard(BaseCardType):
 
 
     @staticmethod
-    def is_custom_font(font: 'Font') -> bool: # type: ignore
+    def is_custom_font(font: 'Font') -> bool:
         """
         Determine whether the given font characteristics constitute a
         default or custom font.
